backend/requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.23.2
psycopg2-binary==2.9.9
pydantic[email]==1.10.13
python-dotenv==1.0.0
email-validator==2.0.0
